requirements.txt
absl-py==2.1.0
appdirs==1.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
gymnasium==0.29.1
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
-e git+https://github.com/VincentStimper/normalizing-flows@848277e2fdd306c16b48dc6814e2d7f8ddaeb2c8#egg=normflows
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
seaborn==0.13.2
sentry-sdk==1.42.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
stable-baselines3==2.2.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.2.1
typing_extensions==4.10.0
tyro==0.7.3
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
Werkzeug==3.0.1
